requirements.txt
Brlapi==0.6.1
Django==1.10.5
Jinja2==2.9.5
Mako==0.9.1
Markdown==2.6.8
MarkupSafe==0.23
PyJWT==1.4.2
Pygments==2.2.0
appdirs==1.4.2
apturl==0.5.2ubuntu4
chardet==2.2.1
checkbox-ng==0.3
checkbox-support==0.2
click==6.7
colorama==0.2.5
command-not-found==0.3
coreapi==2.2.4
coreapi-cli==1.0.6
coreschema==0.0.4
defer==1.0.6
django-cors-headers==2.0.2
django-filter==1.0.1
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
feedparser==5.1.3
friends==0.1
html5lib==0.999
httpie==0.9.9
httplib2==0.8
hunspell==0.1
itypes==1.1.0
language-selector==0.1
louis==2.5.3
lxml==3.3.3
notify2==0.3
oauthlib==0.6.1
onboard==1.0.1
oneconf==0.3.7.14.04.1
packaging==16.8
pbr==2.0.0
piston-mini-client==0.7.5
plainbox==0.5.3
psycopg2==2.7
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.1
pyparsing==2.1.10
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
requests==2.13.0
scudcloud==1.41
six==1.10.0
software-center-aptd-plugins==0.0.0
stevedore==1.20.0
ubuntu-drivers-common==0.0.0
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
unity-scope-audacious==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-clementine==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-gmusicbrowser==0.1
unity-scope-gourmet==0.1
unity-scope-guayadeque==0.1
unity-scope-manpages==0.1
unity-scope-musique==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
uritemplate==3.0.0
urllib3==1.7.1
usb-creator==0.2.23
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wheel==0.24.0
xdiagnose==3.6.3build2
xkit==0.0.0
